requirements.txt
django
djangorestframework
markdown
coreapi
pygments
django-crispy-forms
django-filter
django-tables2
python-dotenv
psycopg2
gunicorn